requirements.txt
flask
gunicorn
python-telegram-bot==20.6
psycopg2-binary
openai
python-dotenv